refactor(chat): modularize ConversationView to pass PR quality baseline (Closes #300)
- Extract SafetyTips component to apps/web/src/components/chat/SafetyTips.tsx
- Extract DateSeparator component to apps/web/src/components/chat/DateSeparator.tsx
- Extract useChatAutoScroll hook to apps/web/src/hooks/useChatAutoScroll.ts
- Reduce ConversationView line count from 361 to 304 lines (passing PR quality threshold)
